use engine initializer to copy javascripts and stylesheets into public dir
i've been happy with the dynamic page caching solution for assets, but it
isn't flexible enough to allow these javascripts and stylesheets to be
added into the main application bundles. it also means extra page loads in
non-caching environments (dev). so i'm moving towards deprecating the
controller-based asset system in favor of the classic copy-assets-
to-public approach.